moved BEC chip into it's own component so it can be used elsewhere
