+++ src/store/index.ts
@@ -30,7 +30,6 @@ import {
   TransportType,
 } from "@/models/newAssessment";
 import createPersistedState from "vuex-persistedstate";
-import { becs } from "./utilities/BECs";
 
 Vue.use(Vuex);
 
@@ -149,9 +148,6 @@ export default new Vuex.Store({
     },
   },
   actions: {
-    getBECs(context) {
-      return becs;
-    },
     // Auth
     logout(context) {
       // wipe the state
