--- src/store/utilities/BECs.ts
@@ -1,37 +1,0 @@
-// Building Element Categories
-
-export interface BEC {
-  name: string;
-  color: string;
-}
-
-export const becs: BEC[] = [
-  {
-    name: "Superstructure",
-    color: "#CD9CE3",
-  },
-  {
-    name: "Substructure",
-    color: "#ED90CB",
-  },
-  {
-    name: "Mechanical Services",
-    color: "#52C4BA",
-  },
-  {
-    name: "Electrical Services",
-    color: "#77DCCC",
-  },
-  {
-    name: "Public Health & Hydraulics",
-    color: "#BFF7EA",
-  },
-  {
-    name: "Space plan",
-    color: "#6CC9F6",
-  },
-  {
-    name: "Building Envelope",
-    color: "#C3EAFE",
-  },
-];
